Add the Cupla CPU backend with synchronous kernel launches
Move the Alpaka and Cupla instructions to a separate file.

Add instructions to build all the backends supported by Cupla, with synchronous
and asynchronous kernel launches.

Rename the Makefile targets to match the Cupla libraries.

Build for the serial CPU backend with synchronous kernel launches; this achieves
the same performance as the "naive" cpu implementation:

  $ ./main-naive
  Got 53376 for cabling, wordCounter 36328
  Output: 1699 modules in 2261.67 us

  $ LD_LIBRARY_PATH=$CUPLA_ROOT/lib ./main-cupla-seq-seq-sync
  Got 53376 for cabling, wordCounter 36328
  Output: 1699 modules in 2252.88 us