Update the instructions to build Cupla as a single header-only library
It is possible to wrap all accelerator-speficic code in Cupla in an inline namespece,
different for each accelerator: see ComputationalRadiationPhysics/cupla#106 .
This leaves the user code unchanged, but allows building Cupla (i.e. linking all
different backends) as a single library.

It is also possible to use Cupla as a header-only library; see
ComputationalRadiationPhysics/cupla#108 for a discussion on the topic, and
ComputationalRadiationPhysics/cupla#109 for a first implementation.