Introduce separate preprocessor symbols for Alpaka client code
ALPAKA_ACC_..._ENABLED is used to build Alpaka code for the accelerator (e.g.
device or kernel code) or the accelerator runtime (e.g. launching a kernel or
device memory operations).
It also sets the new ALPAKA_ACC_..._BACKEND symbol.

ALPAKA_ACC_..._BACKEND is a new symbol that can be used in host code to enable
or call backend-specific functions.