Use experimental printf support in SYCL kernels
According to OpenCL C spec, the printf format string must be in constant
address space.

On the device, define prntf as a macro taking the format as a string
literal and storing it in constant memory, and passes it along with the
other arguments to cl::sycl::intel::experimental::printf.

On the host, use the standard C/C++ implementation.